==> 3_ACP_boletins.Rmd <==
---
title: "ACP Boletins"
author: "Patricia Felipe e Leandro"
date: "11 de junho de 2019"
output: html_document
---

Carregando Pacotes:
```{r echo=FALSE, message=FALSE, warning=FALSE}
library(tidyverse)
library(corrplot)
library(psych)
library(knitr)


library(foreign)
library(psych)
library(dplyr)
library(mvnormtest)
library(GPArotation)

```

Lendo Arquivo:
```{r echo=FALSE}
dados <- read.csv2('base_boletins_tcc.csv', sep=';')

#glimpse(dados)
```

Escalonamento:
```{r message=FALSE, warning=FALSE, paged.print=FALSE}
dados.scale <- scale(dados[,-c(1:2)]) %>% as.data.frame()

```


Teste de esfericidade
```{r echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
bartlett.test(dados)
bartlett.test(dados.scale)
```

Correlações 

```{r echo=FALSE}
#cor(dados.scale)

par(family = 'sans', font=1 )

corrplot.mixed(
    cor(dados.scale), 
    number.cex=0.7,
    lower.col = "darkgrey", 
    tl.cex=0.7,
    tl.col='black'
)

#corrplot(cor(dados.scale), order = "hclust", hclust.method = "ward.D2", addrect = 4)
```


```{r}
# Ajuste do modelo - escolha do numero de componentes 
fit1 <- princomp(dados.scale, cor=TRUE) 
summary(fit1)


```

Gráfico ScreePlot

```{r}
# Scree Plot 
plot(fit1, type="lines", 
    main = 'Gráfico de Cotovelo')
```


Seguindo a Análise com 6 componentes e sem rotação:

```{r message=FALSE, warning=FALSE, paged.print=FALSE}

fit_sem_rotacao <- principal(dados.scale, covar = FALSE, nfactors = 5) 

fit_sem_rotacao


```

Análise de Componentes Principais com rotação Obliqua:

```{r message=FALSE, warning=FALSE, paged.print=FALSE}

fit_oblimin     <- principal(dados.scale, covar = FALSE, nfactors = 5, rotate='oblimin') 
 
fit_oblimin

```

Comparando Cargas Componentes sem rotação e com rotação obliqua:

```{r echo=FALSE}
par(mfrow=c(2,1))
corrplot(t(fit_sem_rotacao$loadings), title = 'Sem Rotação', tl.col="black",tl.cex=0.75 )
corrplot(t(fit_oblimin$loadings) , title = 'Rotação Oblíqua', tl.col="black",tl.cex=0.75)
par(mfrow=c(1,1))
```

Verificando a correlação entre os fatores nos componentes com rotação Oblíqua

```{r echo=FALSE}
cor(fit_oblimin$loadings)
```

Não tem correlação forte, entao  fazer varimax


```{r}
fit_varimax     <- principal(dados.scale, covar = FALSE, nfactors = 5, rotate='varimax') 
 
fit_varimax$loadings
```

```{r}
fit_varimax

```


Comparando as cargas das Rotaçõesw Oblíqua e Varimax

```{r echo=FALSE}

par(mfrow=c(2,1))
corrplot(t(fit_oblimin$loadings), tl.col="black",tl.cex=0.75)
corrplot(t(fit_varimax$loadings), tl.col="black",tl.cex=0.75)
par(mfrow=c(1,1))
```

Relações: (conferir)

RC1:
<br>**X1** - Estupro 
<br>**X2** - Estupro de Vulnerável
<br>**X6** - HOMICÍDIO CULPOSO POR ACIDENTE DE TRÂNSITO
<br>**X11** - LESÃO CORPORAL DOLOSA
<br>**x15** - ROUBO DE CARGA
<br>**x17** - Homicidio doloso exceto transito

RC4:
<br>**X3** - FURTO OUTROS *NEGATIVO*
<br>**X4** - FURTO DE VEÍCULO
<br>**10** - LESÃO CORPORAL CULPOSA POR ACIDENTE DE TRÂNSITO

RC2:
<br>**X9** - LESÃO CORPORAL CULPOSA - OUTRAS
<br>**X14**- ROUBO A BANCO
<br>**X16**- ROUBO DE VEÍCULO *NEGATIVO*

RC3:
<br>**X5** - HOMICÍDIO CULPOSO OUTROS
<br>**X12** - LESÃO CORPORAL SEGUIDA DE MORTE

RC5:
<br>**X7** - HOMICÍDIO DOLOSO POR ACIDENTE DE TRÂNSITO
<br>**X13** - ROUBO - OUTROS



Comunalidades
comunalidade é a variabilidade explicada de cada variavel
(igual nas tres(?))

```{r}
data.frame(
    'Sem Rotação' = fit_sem_rotacao$communality,    
    Oblimin = fit_oblimin$communality,
    Varimax = fit_varimax$communality
)

```

Biplot dos modelos:

```{r}
library(ggfortify)

#autoplot(prcomp(dados.scale))
autoplot(prcomp(dados.scale), data = dados,
         loadings = TRUE, loadings.colour = 'blue',
         loadings.label = TRUE, loadings.label.size = 3)
```



Gravando o arquivo para o próximo passo

```{r echo=FALSE}

base_acp_scores_oblimin <- 
  cbind( dados[,1:2], 
         fit_oblimin$scores)

#head(base_acp_scores_oblimin)

write.csv2(base_acp_scores_oblimin, 'base_acp_scores_oblimin.csv',row.names=FALSE)

```


```{r echo=FALSE}

base_acp_scores_varimax <- 
  cbind( dados[,1:2], 
         fit_varimax$scores)

#head(base_acp_scores_varimax)

write.csv2(base_acp_scores_varimax, 'base_acp_scores_varimax.csv',row.names=FALSE)



```
