==> 5_Descritiva.Rmd <==
---
title: "2.1_Descritiva"
author: "Patricia"
date: "24 de agosto de 2019"
output: html_document
---

Carregando Pacotes:
```{r echo=FALSE, message=FALSE, warning=FALSE}
library(tidyverse)
library(corrplot)
library(psych)
library(knitr)
library(GGally)
library(grid)
library(gridExtra)

```

Lendo Arquivo:
```{r  echo=FALSE}
dados <- read.csv2('base_boletins_tcc.csv', sep=';')

#glimpse(dados)

```


Variáveis
```{r}
variaveis <- read.csv2('variaveis.csv') %>% as.data.frame
print(variaveis)
```

```{r echo=FALSE}
#summary(dados[,-(1:2)])

dados_long <- gather(dados[,-(1:2)], 'var', 'val')
dados_long$var <- dados_long$var %>% as.factor

resumo <- 
  data.frame(
    minimo = tapply(dados_long$val, dados_long$var, min) ,
    maximo = tapply(dados_long$val, dados_long$var, max) ,
    media  = tapply(dados_long$val, dados_long$var, mean),
    mediana  = tapply(dados_long$val, dados_long$var, median), 
    desvio = tapply(dados_long$val, dados_long$var, sd) 
) 


write.csv2(resumo, 'resumo.csv',row.names=TRUE, col.names = TRUE, quote = FALSE)

resumo


```


Distribuição das variáveis originais

```{r echo=FALSE}
par(mfrow=c(2,4))


hist(dados_long[dados_long$var=='X01',2], col = 'lightgrey', main = 'X1', ylab='Frequência', xlab='')
hist(dados_long[dados_long$var=='X02',2], col = 'lightgrey', main = 'X2', ylab='', xlab='')
hist(dados_long[dados_long$var=='X03',2], col = 'lightgrey', main = 'X3', ylab='', xlab='')
hist(dados_long[dados_long$var=='X04',2], col = 'lightgrey', main = 'X4', ylab='', xlab='')
hist(dados_long[dados_long$var=='X05',2], col = 'lightgrey', main = 'X5', ylab='Frequência', xlab='')
hist(dados_long[dados_long$var=='X06',2], col = 'lightgrey', main = 'X6', ylab='', xlab='')
hist(dados_long[dados_long$var=='X07',2], col = 'lightgrey', main = 'X7', ylab='', xlab='')
hist(dados_long[dados_long$var=='X08',2], col = 'lightgrey', main = 'X8', ylab='', xlab='')
hist(dados_long[dados_long$var=='X09',2], col = 'lightgrey', main = 'X9', ylab='Frequência', xlab='')
hist(dados_long[dados_long$var=='X10',2], col = 'lightgrey', main = 'X10', ylab='', xlab='')
hist(dados_long[dados_long$var=='X11',2], col = 'lightgrey', main = 'X11', ylab='', xlab='')
hist(dados_long[dados_long$var=='X12',2], col = 'lightgrey', main = 'X12', ylab='', xlab='')
hist(dados_long[dados_long$var=='X13',2], col = 'lightgrey', main = 'X13', ylab='Frequência', xlab='')
hist(dados_long[dados_long$var=='X14',2], col = 'lightgrey', main = 'X14', ylab='', xlab='')
hist(dados_long[dados_long$var=='X15',2], col = 'lightgrey', main = 'X15', ylab='', xlab='')
hist(dados_long[dados_long$var=='X16',2], col = 'lightgrey', main = 'X16', ylab='', xlab='')


```
```{r include=FALSE}
par(mfrow=c(1,1))

```


Box Plot comparativo das variáveis 

```{r}
par(las = 1)
boxplot(dados_long$val ~ dados_long$var,
        horizontal = T,
        lex.order = TRUE)
```
