==> 6_comparacao_grupos.Rmd <==
---
title: "6_Anova"
author: "Patricia"
date: "1 de setembro de 2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



Importando arquivo dos clusters e das distancias do centro

```{r}

clusters <- read.csv2('result_cluster.csv')

delegacias <- read.csv2('LocDPS.csv')

```


Juntando as duas tabelas

```{r}
library(dplyr)

colnames(delegacias)[1] <- 'delegacia_id'
colnames(delegacias)[2] <- 'Delegacia'
colnames(delegacias)[9] <- 'Região'
colnames(delegacias)[4] <- 'Distância'


dados.dist <- left_join(x= clusters, 
                        y=delegacias, 
                        by='delegacia_id') %>%
               select(delegacia_id, Delegacia, grupos.ward,Distância, 'Região')

```


Distribuição das distâncias
```{r}

library(ggplot2)
Grupos <- as.factor(dados.dist$grupos.ward)
Região <- dados.dist$Região

p <- ggplot(dados.dist, 
              aes(x=Distância, 
                  fill = Grupos)) +
  geom_density(alpha = 0.5)

p

```




Teste de Kruskal-Wallis


The null hypothesis is that the monthly ozone density are identical populations. To test the hypothesis, we apply the kruskal.test function to compare the independent monthly data. The p-value turns out to be nearly zero (6.901e-06). Hence we reject the null hypothesis.

A hipótese nula é de que as distancias do centro são idênticas, então, conforme o p-valor do teste de Kruskal-Wallis é p´roximo de zero, rejeitamos a hipótese nula, aceitando que as médias são diferentes.

```{r}
kruskal.test(Distância ~ grupos.ward, data = dados.dist) 
 

```


```{r}
pairwise.wilcox.test(x= dados.dist$Distância, 
                     g= Grupos, 
                     p.adj = "bonf",
                     paired = FALSE)

```




Box Plots
```{r}

p <- ggplot(dados.dist, aes(x=Grupos, y=Distância, fill=Grupos, alpha=0.7)) + 
  geom_boxplot()+
  ggtitle('Variação das Distâncias do Centro por Grupo')
p

```


Outra forma de calcular as diferencas entre as medias

```{r}

library(ggpubr)
#theme_set(theme_pubclean())



compare_means( Distância ~  grupos.ward, 
               data = dados.dist,
               method = "kruskal.test")

```



```{r}

mycomp <- list(c('1','2'), 
               c('1','3'),
               c('1','4'),
               c('1','5'),
               c('2','3'),
               c('2','4'),
               c('2','5'),
               c('3','4'),
               c('3','5'),
               c('4','5')
               )

ggboxplot(dados.dist, 
          x = 'grupos.ward',
          y = 'Distância',
          fill = "grupos.ward",
          alpha =0.8) + 
  stat_compare_means(comparisons = mycomp) + 
  stat_compare_means(label.y = 70)            



```

Grupo 1 (a, b);
Grupo 2 (a, c);
Grupo 3 (a, b);
Grupo 4 (c);
Grupo 5 (b).

```{r}

rotulos <- dados.dist$grupos.ward
rotulos[rotulos==1] <- 'a , b'
rotulos[rotulos==2] <- 'a , c'
rotulos[rotulos==3] <- 'a , b'
rotulos[rotulos==4] <- 'c'
rotulos[rotulos==5] <- 'b'


ggplot(data=dados.dist,
    aes(x = grupos.ward,
    y = Distância)
    ) + 
    geom_boxplot(aes(fill=Grupos, 
                     alpha=0.8,
                     labels='asdf')) + 
  geom_text(data = dados.dist, 
            aes(x = grupos.ward, 
                y = 30, 
                label = rotulos), 
            vjust = -0.5) # use hjust to adjust horizontal position 
  
    
```



##TESTE DE DUNN
```{r}

library(FSA)
# dunnTest(x= dados.dist$Distância, g= Grupos)
 
dunn_result <- dunnTest(x= dados.dist$Distância, 
                     g= Grupos)

print(dunn_result,dunn.test.results=FALSE)

print(dunn_result,dunn.test.results=TRUE)
```
                     
## grupos teste de dunns kruskall wallis multiple comparsions

```{r message=FALSE, warning=FALSE, paged.print=FALSE}

rotulos <- dados.dist$grupos.ward
rotulos[rotulos==1] <- 'a , b, c'
rotulos[rotulos==2] <- 'a , b'
rotulos[rotulos==3] <- 'a , c'
rotulos[rotulos==4] <- 'b'
rotulos[rotulos==5] <- 'c'


ggplot(data=dados.dist,
    aes(x = grupos.ward,
    y = Distância)
    ) + 
    geom_boxplot(aes(fill=Grupos, 
                     alpha=0.8,
                     labels='asdf')) + 
  geom_text(data = dados.dist, 
            aes(x = grupos.ward, 
                y = 30, 
                label = rotulos), 
            vjust = -0.5) # use hjust to adjust horizontal position 
  
    
```
```
